Refs #IUS-302 ~ Configuration of migration 0.0.6 updated

+++ Migrations/pdo_pgsql/Version006.php
@@ -9,6 +9,15 @@ class Version006 extends AbstractMigration
 {
     const VERSION = '0.0.6';
 
+    private function tableExists($table, $schema = 'public')
+    {
+        $statement = $this->connection->prepare("SELECT EXISTS(SELECT 1 FROM pg_catalog.pg_class c JOIN pg_catalog.pg_namespace n ON n.oid = c.relnamespace WHERE n.nspname = '" . $schema . "' AND c.relname = '" . $table . "');");
+        $statement->execute();
+        $result = $statement->fetchAll();
+
+        return ($result[0]['exists']);
+    }
+
     public function getName()
     {
         return self::VERSION;
@@ -16,36 +25,38 @@ class Version006 extends AbstractMigration
 
     public function up(Schema $schema)
     {
-        $this->addSql('DROP TABLE public.migration_versions;');
+        if ($this->tableExists('migration_versions')) {
+            $this->addSql('DROP TABLE public.migration_versions;');
 
-        $this->addSql('CREATE TABLE public.doctrine_canaltpmatrixbundle_versions (version character varying(255) NOT NULL);');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmatrixbundle_versions (version) VALUES (\'0.0.1\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmatrixbundle_versions (version) VALUES (\'0.0.2\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmatrixbundle_versions (version) VALUES (\'0.0.3\');');
+            $this->addSql('CREATE TABLE public.doctrine_canaltpmatrixbundle_versions (version character varying(255) NOT NULL);');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmatrixbundle_versions (version) VALUES (\'0.0.1\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmatrixbundle_versions (version) VALUES (\'0.0.2\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmatrixbundle_versions (version) VALUES (\'0.0.3\');');
 
-        $this->addSql('CREATE TABLE public.doctrine_canaltpmttbundle_versions (version character varying(255) NOT NULL);');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.1\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.2\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.3\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.4\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.5\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.6\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.7\');');
+            $this->addSql('CREATE TABLE public.doctrine_canaltpmttbundle_versions (version character varying(255) NOT NULL);');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.1\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.2\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.3\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.4\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.5\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.6\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpmttbundle_versions (version) VALUES (\'0.0.7\');');
 
-        $this->addSql('CREATE TABLE public.doctrine_canaltpnmpadminbundle_versions (version character varying(255) NOT NULL);');
-        $this->addSql('INSERT INTO public.doctrine_canaltpnmpadminbundle_versions (version) VALUES (\'0.0.1\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpnmpadminbundle_versions (version) VALUES (\'0.0.2\');');
+            $this->addSql('CREATE TABLE public.doctrine_canaltpnmpadminbundle_versions (version character varying(255) NOT NULL);');
+            $this->addSql('INSERT INTO public.doctrine_canaltpnmpadminbundle_versions (version) VALUES (\'0.0.1\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpnmpadminbundle_versions (version) VALUES (\'0.0.2\');');
 
-        $this->addSql('CREATE TABLE public.doctrine_canaltprealtimebundle_versions (version character varying(255) NOT NULL);');
-        $this->addSql('INSERT INTO public.doctrine_canaltprealtimebundle_versions (version) VALUES (\'0.0.1\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltprealtimebundle_versions (version) VALUES (\'0.0.2\');');
+            $this->addSql('CREATE TABLE public.doctrine_canaltprealtimebundle_versions (version character varying(255) NOT NULL);');
+            $this->addSql('INSERT INTO public.doctrine_canaltprealtimebundle_versions (version) VALUES (\'0.0.1\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltprealtimebundle_versions (version) VALUES (\'0.0.2\');');
 
-        $this->addSql('CREATE TABLE public.doctrine_canaltpsamcorebundle_versions (version character varying(255) NOT NULL);');
-        $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.1\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.2\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.3\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.4\');');
-        $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.5\');');
+            $this->addSql('CREATE TABLE public.doctrine_canaltpsamcorebundle_versions (version character varying(255) NOT NULL);');
+            $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.1\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.2\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.3\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.4\');');
+            $this->addSql('INSERT INTO public.doctrine_canaltpsamcorebundle_versions (version) VALUES (\'0.0.5\');');
+        }
     }
 
     public function down(Schema $schema)
